% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1hdexplorer.R
\name{load.HDpackages}
\alias{load.HDpackages}
\title{Loads Packages}
\usage{
load.HDpackages(some.packages, all.packages)
}
\arguments{
\item{some.packages}{Loads only required packages. These include: corrplot,dendextend,factoextra,FactoMineR,ggplot2,gplots,Hmisc,psych,RColorBrewer,RSKC,Rtsne,seriation,sparcl,stats,tsne, vioplot.}

\item{all.packages}{Loads all packages contained in 'some.packages' as well as additional recommended packages. Additional packages include: autoimage,data.table,d3heatmap,dendsort,GGally,ggpubr,ggrepel,gridExtra,miscTools,plotrix,plyr,reshape, scales}
}
\value{
This will consult your library and load the desired set of packages (some.packages or all.packages).
}
\description{
Loads Packages
}
\examples{
load.HDpackages(some.packages=TRUE) #Will load the packages in some.packages to your library
}
