% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1hdexplorer.R
\name{install.HDpackages}
\alias{install.HDpackages}
\title{Installs Packages}
\usage{
install.HDpackages(all.packages, some.packages)
}
\arguments{
\item{all.packages}{Installs all available packages}

\item{some.packages}{Installs only required packages}
}
\description{
Installs Packages
}
