% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1hdexplorer.R
\name{load.HDpackages}
\alias{load.HDpackages}
\title{Loads Packages}
\usage{
load.HDpackages(all.packages, some.packages)
}
\arguments{
\item{all.packages}{Loads all available packages}

\item{some.packages}{Loads only required packages}
}
\description{
Loads Packages
}
